% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{CancelJobRequest}
\alias{CancelJobRequest}
\title{CancelJobRequest Object}
\usage{
CancelJobRequest()
}
\value{
CancelJobRequest object
}
\description{
CancelJobRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request to cancel a job.
}
\seealso{
Other CancelJobRequest functions: \code{\link{projects.regions.jobs.cancel}}
}

