% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{ClusterConfiguration}
\alias{ClusterConfiguration}
\title{ClusterConfiguration Object}
\usage{
ClusterConfiguration(configurationBucket = NULL, gceConfiguration = NULL,
  numWorkers = NULL, workers = NULL, masterName = NULL,
  masterDiskConfiguration = NULL, workerDiskConfiguration = NULL,
  gceClusterConfiguration = NULL, masterConfiguration = NULL,
  workerConfiguration = NULL, secondaryWorkerConfiguration = NULL,
  softwareConfiguration = NULL, initializationActions = NULL)
}
\arguments{
\item{configurationBucket}{[Optional] A Google Cloud Storage staging bucket used for sharing generated SSH keys and configuration}

\item{gceConfiguration}{[Deprecated] The Google Compute Engine configuration settings for cluster resources}

\item{numWorkers}{[Deprecated] The number of worker nodes in the cluster}

\item{workers}{[Deprecated] The list of worker node names}

\item{masterName}{[Deprecated] The Master's hostname}

\item{masterDiskConfiguration}{[Deprecated] The configuration settings of master node disk options}

\item{workerDiskConfiguration}{[Deprecated] The configuration settings of worker node disk options}

\item{gceClusterConfiguration}{[Optional] The shared Google Compute Engine configuration settings for all instances in a cluster}

\item{masterConfiguration}{[Optional] The Google Compute Engine configuration settings for the master instance in a cluster}

\item{workerConfiguration}{[Optional] The Google Compute Engine configuration settings for worker instances in a cluster}

\item{secondaryWorkerConfiguration}{[Optional] The Google Compute Engine configuration settings for additional worker instances in a cluster}

\item{softwareConfiguration}{[Optional] The configuration settings for software inside the cluster}

\item{initializationActions}{[Optional] Commands to execute on each node after configuration is completed}
}
\value{
ClusterConfiguration object
}
\description{
ClusterConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The cluster configuration.
}

