% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{ClusterMetrics}
\alias{ClusterMetrics}
\title{ClusterMetrics Object}
\usage{
ClusterMetrics(ClusterMetrics.hdfsMetrics = NULL,
  ClusterMetrics.yarnMetrics = NULL, hdfsMetrics = NULL,
  yarnMetrics = NULL)
}
\arguments{
\item{ClusterMetrics.hdfsMetrics}{The \link{ClusterMetrics.hdfsMetrics} object or list of objects}

\item{ClusterMetrics.yarnMetrics}{The \link{ClusterMetrics.yarnMetrics} object or list of objects}

\item{hdfsMetrics}{The HDFS metrics}

\item{yarnMetrics}{The YARN metrics}
}
\value{
ClusterMetrics object
}
\description{
ClusterMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains cluster daemon metrics, such as HDFS and YARN stats.
}
\seealso{
Other ClusterMetrics functions: \code{\link{ClusterMetrics.hdfsMetrics}},
  \code{\link{ClusterMetrics.yarnMetrics}}
}
