% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{ClusterOperationStatus}
\alias{ClusterOperationStatus}
\title{ClusterOperationStatus Object}
\usage{
ClusterOperationStatus(state = NULL, details = NULL, innerState = NULL,
  stateStartTime = NULL)
}
\arguments{
\item{state}{Output-only A message containing the operation state}

\item{details}{Output-onlyA message containing any operation metadata details}

\item{innerState}{Output-only A message containing the detailed operation state}

\item{stateStartTime}{Output-only The time this state was entered}
}
\value{
ClusterOperationStatus object
}
\description{
ClusterOperationStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The status of the operation.
}
