% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{ClusterStatus}
\alias{ClusterStatus}
\title{ClusterStatus Object}
\usage{
ClusterStatus(state = NULL, detail = NULL, stateStartTime = NULL)
}
\arguments{
\item{state}{The cluster's state}

\item{detail}{Optional details of cluster's state}

\item{stateStartTime}{Time when this state was entered}
}
\value{
ClusterStatus object
}
\description{
ClusterStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The status of a cluster and its instances.
}

