% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_functions.R
\docType{package}
\name{dataproc_googleAuthR}
\alias{dataproc_googleAuthR}
\alias{dataproc_googleAuthR-package}
\title{Google Cloud Dataproc API
Manages Hadoop-based clusters and jobs on Google Cloud Platform.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:42:12
filename: /Users/mark/dev/R/autoGoogleAPI/googledataprocv1alpha1.auto/R/dataproc_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}
}
