% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{GceClusterConfiguration.metadata}
\alias{GceClusterConfiguration.metadata}
\title{GceClusterConfiguration.metadata Object}
\usage{
GceClusterConfiguration.metadata()
}
\value{
GceClusterConfiguration.metadata object
}
\description{
GceClusterConfiguration.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The Google Compute Engine metadata entries to add to all instances.
}
\seealso{
Other GceClusterConfiguration functions: \code{\link{GceClusterConfiguration}}
}
