% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{GceClusterConfiguration}
\alias{GceClusterConfiguration}
\title{Google Cloud Dataproc API Objects 
Manages Hadoop-based clusters and jobs on Google Cloud Platform.}
\usage{
GceClusterConfiguration(GceClusterConfiguration.metadata = NULL,
  metadata = NULL, internalIpOnly = NULL, serviceAccountScopes = NULL,
  tags = NULL, serviceAccount = NULL, subnetworkUri = NULL,
  networkUri = NULL, zoneUri = NULL)
}
\arguments{
\item{GceClusterConfiguration.metadata}{The \link{GceClusterConfiguration.metadata} object or list of objects}

\item{metadata}{The Google Compute Engine metadata entries to add to all instances}

\item{internalIpOnly}{If true, all instances in the cluser will only have internal IP addresses}

\item{serviceAccountScopes}{The service account scopes included in Google Compute Engine instances}

\item{tags}{The Google Compute Engine tags to add to all instances}

\item{serviceAccount}{Optional The service account of the instances}

\item{subnetworkUri}{The Google Compute Engine subnetwork to be used for machine communications}

\item{networkUri}{The Google Compute Engine network to be used for machine communications}

\item{zoneUri}{Required The zone where the Google Compute Engine cluster will be located}
}
\value{
GceClusterConfiguration object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:42:12
filename: /Users/mark/dev/R/autoGoogleAPI/googledataprocv1alpha1.auto/R/dataproc_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
GceClusterConfiguration Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Common configuration settings for resources of Google Compute Engine cluster instances, applicable to all instances in the cluster.
}
\seealso{
Other GceClusterConfiguration functions: \code{\link{GceClusterConfiguration.metadata}}
}
