% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{GceConfiguration}
\alias{GceConfiguration}
\title{GceConfiguration Object}
\usage{
GceConfiguration(networkUri = NULL, serviceAccountScopes = NULL,
  zoneUri = NULL, imageUri = NULL, machineTypeUri = NULL)
}
\arguments{
\item{networkUri}{Deprecated The Google Compute Engine network to be used for machine communications}

\item{serviceAccountScopes}{Deprecated The service account scopes included in Google Compute Engine instances}

\item{zoneUri}{Deprecated The zone where the Google Compute Engine cluster will be located}

\item{imageUri}{Deprecated The Google Compute Engine image resource used for cluster instances}

\item{machineTypeUri}{Deprecated The Google Compute Engine machine type used for cluster instances}
}
\value{
GceConfiguration object
}
\description{
GceConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Deprecated Common configuration settings for resources of Google Compute Engine cluster instances, applicable to all instances in the cluster.
}
