% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{HadoopJob.properties}
\alias{HadoopJob.properties}
\title{HadoopJob.properties Object}
\usage{
HadoopJob.properties()
}
\value{
HadoopJob.properties object
}
\description{
HadoopJob.properties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional A mapping of property names to values, used to configure Hadoop. Properties that conflict with values set by the Cloud Dataproc API may be overwritten. Can include properties set in /etc/hadoop/conf/*-site and classes in user code.
}
\seealso{
Other HadoopJob functions: \code{\link{HadoopJob}}
}
