% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{HiveJob.scriptVariables}
\alias{HiveJob.scriptVariables}
\title{HiveJob.scriptVariables Object}
\usage{
HiveJob.scriptVariables()
}
\value{
HiveJob.scriptVariables object
}
\description{
HiveJob.scriptVariables Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Optional] Mapping of query variable names to values (equivalent to the Hive command: 'SET name='value';').
}
\seealso{
Other HiveJob functions: \code{\link{HiveJob.properties}},
  \code{\link{HiveJob}}
}

