% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{InstanceGroupConfiguration}
\alias{InstanceGroupConfiguration}
\title{InstanceGroupConfiguration Object}
\usage{
InstanceGroupConfiguration(numInstances = NULL, instanceNames = NULL,
  imageUri = NULL, machineTypeUri = NULL, diskConfiguration = NULL,
  isPreemptible = NULL, managedGroupConfiguration = NULL)
}
\arguments{
\item{numInstances}{The number of VM instances in the instance group}

\item{instanceNames}{The list of instance names}

\item{imageUri}{[Output-only] The Google Compute Engine image resource used for cluster instances}

\item{machineTypeUri}{The Google Compute Engine machine type used for cluster instances}

\item{diskConfiguration}{Disk option configuration settings}

\item{isPreemptible}{Specifies that this instance group contains Preemptible Instances}

\item{managedGroupConfiguration}{[Output-only] The configuration for Google Compute Engine Instance Group Manager that manages this group}
}
\value{
InstanceGroupConfiguration object
}
\description{
InstanceGroupConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The configuration settings for Google Compute Engine resources in an instance group, such as a master or worker group.
}

