% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{Job.labels}
\alias{Job.labels}
\title{Job.labels Object}
\usage{
Job.labels()
}
\value{
Job.labels object
}
\description{
Job.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional The labels to associate with this job.Label keys must be between 1 and 63 characters long, and must conform to the following regular expression: \p{Ll}\p{Lo}{0,62}Label values must be between 1 and 63 characters long, and must conform to the following regular expression: \p{Ll}\p{Lo}\p{N}_-{0,63}No more than 64 labels can be associated with a given job.
}
\seealso{
Other Job functions: \code{\link{Job}},
  \code{\link{projects.regions.jobs.patch}}
}
