% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{JobStatus}
\alias{JobStatus}
\title{JobStatus Object}
\usage{
JobStatus(startTime = NULL, insertTime = NULL, state = NULL,
  details = NULL, stateStartTime = NULL, endTime = NULL)
}
\arguments{
\item{startTime}{The time when the server started the job}

\item{insertTime}{The time of the job request}

\item{state}{Required A state message specifying the overall job state}

\item{details}{Optional Job state details, such as an error description if the state is <code>ERROR</code>}

\item{stateStartTime}{Output-only The time when this state was entered}

\item{endTime}{The time when the job completed}
}
\value{
JobStatus object
}
\description{
JobStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Dataproc job status.
}
