% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{ListJobsResponse}
\alias{ListJobsResponse}
\title{ListJobsResponse Object}
\usage{
ListJobsResponse(nextPageToken = NULL, jobs = NULL)
}
\arguments{
\item{nextPageToken}{Optional This token is included in the response if there are more results to fetch}

\item{jobs}{Output-only Jobs list}
}
\value{
ListJobsResponse object
}
\description{
ListJobsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A response to a request to list jobs in a project.
}
