% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{ManagedGroupConfiguration}
\alias{ManagedGroupConfiguration}
\title{ManagedGroupConfiguration Object}
\usage{
ManagedGroupConfiguration(instanceGroupManagerName = NULL,
  instanceTemplateName = NULL)
}
\arguments{
\item{instanceGroupManagerName}{Output-only The name of Instance Group Manager managing this group}

\item{instanceTemplateName}{Output-only The name of Instance Template used for Managed Instance Group}
}
\value{
ManagedGroupConfiguration object
}
\description{
ManagedGroupConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies the resources used to actively manage an instance group.
}
