% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{Operation.metadata}
\alias{Operation.metadata}
\title{Operation.metadata Object}
\usage{
Operation.metadata()
}
\value{
Operation.metadata object
}
\description{
Operation.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Service-specific metadata associated with the operation.
}
\seealso{
Other Operation functions: \code{\link{Operation.response}},
  \code{\link{Operation}}
}

