% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{Operation.response}
\alias{Operation.response}
\title{Operation.response Object}
\usage{
Operation.response()
}
\value{
Operation.response object
}
\description{
Operation.response Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The operation response. If the called method returns no data on success, the response is google.protobuf.Empty. If the called method is Get,Create or Update, the response is the resource. For all other methods, the response type is a concatenation of the method name and 'Response'. For example, if the called method is TakeSnapshot(), the response type is TakeSnapshotResponse.
}
\seealso{
Other Operation functions: \code{\link{Operation.metadata}},
  \code{\link{Operation}}
}
