% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{OperationMetadata}
\alias{OperationMetadata}
\title{OperationMetadata Object}
\usage{
OperationMetadata(state = NULL, innerState = NULL, details = NULL,
  insertTime = NULL, startTime = NULL, endTime = NULL,
  clusterName = NULL, clusterUuid = NULL, status = NULL,
  statusHistory = NULL, operationType = NULL, description = NULL)
}
\arguments{
\item{state}{A message containing the operation state}

\item{innerState}{A message containing the detailed operation state}

\item{details}{A message containing any operation metadata details}

\item{insertTime}{The time that the operation was requested}

\item{startTime}{The time that the operation was started by the server}

\item{endTime}{The time that the operation completed}

\item{clusterName}{Name of the cluster for the operation}

\item{clusterUuid}{Cluster UUId for the operation}

\item{status}{[Output-only] Current operation status}

\item{statusHistory}{[Output-only] Previous operation status}

\item{operationType}{[Output-only] The operation type}

\item{description}{[Output-only] Short description of operation}
}
\value{
OperationMetadata object
}
\description{
OperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata describing the operation.
}

