% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{PigJob.scriptVariables}
\alias{PigJob.scriptVariables}
\title{PigJob.scriptVariables Object}
\usage{
PigJob.scriptVariables()
}
\value{
PigJob.scriptVariables object
}
\description{
PigJob.scriptVariables Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional Mapping of query variable names to values (equivalent to the Pig command: 'name=value').
}
\seealso{
Other PigJob functions: \code{\link{PigJob.properties}},
  \code{\link{PigJob}}
}
