% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{PySparkJob.properties}
\alias{PySparkJob.properties}
\title{PySparkJob.properties Object}
\usage{
PySparkJob.properties()
}
\value{
PySparkJob.properties object
}
\description{
PySparkJob.properties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional A mapping of property names to values, used to configure PySpark. Properties that conflict with values set by the Cloud Dataproc API may be overwritten. Can include properties set in /etc/spark/conf/spark-defaults.conf and classes in user code.
}
\seealso{
Other PySparkJob functions: \code{\link{PySparkJob}}
}
