% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{SparkSqlJob}
\alias{SparkSqlJob}
\title{SparkSqlJob Object}
\usage{
SparkSqlJob(SparkSqlJob.scriptVariables = NULL,
  SparkSqlJob.properties = NULL, queryFileUri = NULL, queryList = NULL,
  scriptVariables = NULL, properties = NULL, jarFileUris = NULL,
  loggingConfiguration = NULL)
}
\arguments{
\item{SparkSqlJob.scriptVariables}{The \link{SparkSqlJob.scriptVariables} object or list of objects}

\item{SparkSqlJob.properties}{The \link{SparkSqlJob.properties} object or list of objects}

\item{queryFileUri}{The HCFS URI of the script that contains SQL queries}

\item{queryList}{A list of queries}

\item{scriptVariables}{[Optional] Mapping of query variable names to values (equivalent to the Spark SQL command: SET name='value';)}

\item{properties}{[Optional] A mapping of property names to values, used to configure Spark SQL's SparkConf}

\item{jarFileUris}{[Optional] HCFS URIs of jar files to be added to the Spark CLASSPATH}

\item{loggingConfiguration}{[Optional] The runtime log configuration for job execution}
}
\value{
SparkSqlJob object
}
\description{
SparkSqlJob Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Cloud Dataproc job for running Spark SQL queries.
}
\seealso{
Other SparkSqlJob functions: \code{\link{SparkSqlJob.properties}},
  \code{\link{SparkSqlJob.scriptVariables}}
}

