% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{YarnApplication}
\alias{YarnApplication}
\title{YarnApplication Object}
\usage{
YarnApplication(name = NULL, trackingUrl = NULL, progress = NULL,
  state = NULL)
}
\arguments{
\item{name}{Required The application name}

\item{trackingUrl}{Optional The HTTP URL of the ApplicationMaster, HistoryServer, or TimelineServer that provides application-specific information}

\item{progress}{Required The numerical progress of the application, from 1 to 100}

\item{state}{Required The application state}
}
\value{
YarnApplication object
}
\description{
YarnApplication Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A YARN application created by a job. Application information is a subset of <code>org.apache.hadoop.yarn.proto.YarnProtos.ApplicationReportProto</code>.
}
