% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{ClusterConfiguration}
\alias{ClusterConfiguration}
\title{ClusterConfiguration Object}
\usage{
ClusterConfiguration(workerDiskConfiguration = NULL,
  gceConfiguration = NULL, masterConfiguration = NULL,
  secondaryWorkerConfiguration = NULL, masterName = NULL, workers = NULL,
  workerConfiguration = NULL, initializationActions = NULL,
  softwareConfiguration = NULL, gceClusterConfiguration = NULL,
  configurationBucket = NULL, numWorkers = NULL,
  masterDiskConfiguration = NULL)
}
\arguments{
\item{workerDiskConfiguration}{Deprecated The configuration settings of worker node disk options}

\item{gceConfiguration}{Deprecated The Google Compute Engine configuration settings for cluster resources}

\item{masterConfiguration}{Optional The Google Compute Engine configuration settings for the master instance in a cluster}

\item{secondaryWorkerConfiguration}{Optional The Google Compute Engine configuration settings for additional worker instances in a cluster}

\item{masterName}{Deprecated The Master's hostname}

\item{workers}{Deprecated The list of worker node names}

\item{workerConfiguration}{Optional The Google Compute Engine configuration settings for worker instances in a cluster}

\item{initializationActions}{Optional Commands to execute on each node after configuration is completed}

\item{softwareConfiguration}{Optional The configuration settings for software inside the cluster}

\item{gceClusterConfiguration}{Optional The shared Google Compute Engine configuration settings for all instances in a cluster}

\item{configurationBucket}{Optional A Google Cloud Storage staging bucket used for sharing generated SSH keys and configuration}

\item{numWorkers}{Deprecated The number of worker nodes in the cluster}

\item{masterDiskConfiguration}{Deprecated The configuration settings of master node disk options}
}
\value{
ClusterConfiguration object
}
\description{
ClusterConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The cluster configuration.
}
