% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{GceClusterConfiguration}
\alias{GceClusterConfiguration}
\title{GceClusterConfiguration Object}
\usage{
GceClusterConfiguration(GceClusterConfiguration.metadata = NULL,
  zoneUri = NULL, networkUri = NULL, subnetworkUri = NULL,
  serviceAccountScopes = NULL, tags = NULL, metadata = NULL)
}
\arguments{
\item{GceClusterConfiguration.metadata}{The \link{GceClusterConfiguration.metadata} object or list of objects}

\item{zoneUri}{[Required] The zone where the Google Compute Engine cluster will be located}

\item{networkUri}{The Google Compute Engine network to be used for machine communications}

\item{subnetworkUri}{The Google Compute Engine subnetwork to be used for machine communications}

\item{serviceAccountScopes}{The service account scopes included in Google Compute Engine instances}

\item{tags}{The Google Compute Engine tags to add to all instances}

\item{metadata}{The Google Compute Engine metadata entries to add to all instances}
}
\value{
GceClusterConfiguration object
}
\description{
GceClusterConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Common configuration settings for resources of Google Compute Engine cluster instances, applicable to all instances in the cluster.
}
\seealso{
Other GceClusterConfiguration functions: \code{\link{GceClusterConfiguration.metadata}}
}

