% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{ListJobsRequest}
\alias{ListJobsRequest}
\title{ListJobsRequest Object}
\usage{
ListJobsRequest(pageSize = NULL, filter = NULL, jobStateMatcher = NULL,
  pageToken = NULL, clusterName = NULL)
}
\arguments{
\item{pageSize}{Optional The number of results to return in each response}

\item{filter}{Optional A filter constraining which jobs to list}

\item{jobStateMatcher}{Optional Specifies enumerated categories of jobs to list}

\item{pageToken}{Optional The page token, returned by a previous call, to request the next page of results}

\item{clusterName}{Optional If set, the returned jobs list includes only jobs that were submitted to the named cluster}
}
\value{
ListJobsRequest object
}
\description{
ListJobsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request to list jobs in a project.
}
\seealso{
Other ListJobsRequest functions: \code{\link{projects.regions.jobs.list}}
}
